# fix: performance improvements when meta_generator is disabled